requirements.txt
asn1crypto==0.24.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.3.1
Django==2.1.1
django-crispy-forms==1.7.2
django-filter==2.0.0
django-guardian==1.4.9
django-rest-knox==3.1.5
djangorestframework==3.8.2
idna==2.7
itypes==1.1.0
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.0
numpy==1.15.2
pandas==0.23.4
pycparser==2.19
PyJWT==1.6.4
pyOpenSSL==18.0.0
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
